Move FilterIterator
